+++ src/Laravel/Facades/KonturFocus.php
@@ -18,7 +18,20 @@ use Ex3mm\KonturFocus\Contracts\RequestBuilderInterface;
  * @method static RequestBuilderInterface linkAnalytics()
  * @method static RequestBuilderInterface purchasesAnalytics()
  * @method static RequestBuilderInterface licenses()
+ * @method static RequestBuilderInterface beneficiary()
  * @method static RequestBuilderInterface custom(string $endpoint)
+ *
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getReq(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getEgrDetails(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getLegalAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getBankruptcyAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getCourtAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getFinanceAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getFsspAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getLinkAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getPurchasesAnalytics(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getLicenses(?string $inn = null, ?string $ogrn = null)
+ * @method static \Ex3mm\KonturFocus\DTOs\CollectionResponse getBeneficiary(?string $inn = null, ?string $ogrn = null)
  *
  * @see \Ex3mm\KonturFocus\KonturFocusManager
  */